feat(database): implement drag-and-drop reordering for database views
- add composable to manage view reordering with persistence
- integrate drag lifecycle handling and external update reconciliation
- ensure correct behavior during concurrent updates and rejections
- cover functionality with comprehensive unit tests
